Move `ConcatSlice` to util
